Renamed some mock/local files and ignored some
Some modules which are used for local development only, were bringing
our coverage numbers down.  It was determined that since they are used
only for development in a local-only capacity, that tracking testing
coverage for them was less important.  Therefore we have added those
files in particular to the jest config to ignore them for coverage.

Other files were renamed to make their use more clear.

branch coverage is now at 90% on the back end.

Jira ticket: CAMS-249
