+++ backend/functions/jest.config.js
@@ -7,23 +7,21 @@ module.exports = {
   coverageThreshold: {
     global: {
       lines: 90,
-      branches: 85,
+      branches: 90,
     },
   },
   coveragePathIgnorePatterns: [
     '<rootDir>/node_modules/',
     '.dependency-cruiser.js',
     '.*test.[jt]s',
+    '.*mock.*.ts',
+    '.*.d.ts',
     'azure/app-insights.ts',
     'dist/',
     'coverage/',
-    'lib/interfaces/case.assignment.repository.interface.d.ts',
-    'lib/use-cases/cases.interface.d.ts',
-    'lib/use-cases/attorney.gateway.interface.d.ts',
-    'lib/adapters/types/',
-    'lib/adapters/gateways/cases.local.gateway.ts',
-    'lib/testing/',
     'lib/cosmos-humble-objects/',
+    'lib/testing/mock-data/index.ts',
+    'lib/testing/testing-utilities.ts',
     'jest.*config.js',
   ],
   testMatch: ['**/?(*.)+(spec|test).[jt]s?(x)', '!**/?(*.)+(integration).(spec|test).[jt]s?(x)'],
